requirements.txt
pandas==2.0.3
numpy==1.24.3
openai==0.28.0
python-dotenv==1.0.0
networkx==3.1
matplotlib==3.7.1
streamlit==1.35.0
python==3.11.9